requirements.txt
numpy>=1.19.0
matplotlib>=3.3.0
opencv-python>=4.5.0
pytesseract>=0.3.0
Pillow>=8.0.0
sympy>=1.8.0
